requirements.txt
click>=8.1.7
pyYaml>=6.0.1
cffi>=1.16.0
beautifulsoup4>=4.12.3
alive-progress>=3.1.5
